move trusted sites function logic to shared helper func
