+++ components/d2l-activity-editor/d2l-activity-assignment-editor/d2l-activity-assignment-editor.js
@@ -13,6 +13,7 @@ import { LocalizeActivityAssignmentEditorMixin } from './mixins/d2l-activity-ass
 import { MobxLitElement } from '@adobe/lit-mobx';
 import { RtlMixin } from '@brightspace-ui/core/mixins/rtl-mixin.js';
 import { shared as store } from '../state/activity-store.js';
+import { trustedSitesProviderFn } from '../shared/trusted-site-provider.js';
 
 class AssignmentEditor extends AsyncContainerMixin(RtlMixin(LocalizeActivityAssignmentEditorMixin(ActivityEditorMixin(MobxLitElement)))) {
 
@@ -220,39 +221,7 @@ class AssignmentEditor extends AsyncContainerMixin(RtlMixin(LocalizeActivityAssi
 		// Provides function to validate if a URL is trusted for d2l-labs-attachment
 		// https://github.com/Brightspace/attachment/blob/e44cab1f0cecc55dd93acf59212fabc6872c0bd3/components/attachment.js#L115
 		if (e.detail.key === 'd2l-provider-trusted-site-fn') {
-			e.detail.provider = () => url => {
-				const origin = new URL(url).origin;
-				const unfilteredContent = `<iframe src="${origin}"></iframe>`;
-
-				return new Promise((resolve, reject) => {
-					const params = {
-						filterMode: 1, // strict mode for html filtering. Refer to D2L.LP.TextProcessing.FilterModes
-						html: unfilteredContent
-					};
-					const options = {
-						success: resolve,
-						failure: reject
-					};
-					D2L.LP.Web.UI.Rpc.Connect(
-						D2L.LP.Web.UI.Rpc.Verbs.POST,
-						new D2L.LP.Web.Http.UrlLocation(this.trustedSitesEndpoint),
-						params,
-						options
-					);
-				}).then(filteredContent => {
-					const matchSrc = function(str) {
-						// excludes matching query string as filterHtml may modify the query string
-						return str.match(/src=["']([^?"']+)/i);
-					};
-					const unfilteredMatch = matchSrc(unfilteredContent);
-					const unfilteredSrc = unfilteredMatch && unfilteredMatch.length === 2 && unfilteredMatch[1];
-
-					const filteredMatch = matchSrc(filteredContent);
-					const filteredSrc = filteredMatch && filteredMatch.length === 2 && filteredMatch[1];
-
-					return unfilteredSrc === filteredSrc;
-				});
-			};
+			e.detail.provider = trustedSitesProviderFn(this.trustedSitesEndpoint);
 			e.stopPropagation();
 		}
 	}
